# api/openapi.yaml
openapi: 3.1.0
info:
  title: coc-pointer API
  version: 0.6.0
  description: |
    클래시 오브 클랜 클랜 "미니언즈"의 클랜원과 점수를 다루는 API.

    이 파일이 API 스펙이다. 서버 모델과 프론트 타입은 여기서 생성하며
    손으로 고치지 않는다. scripts/generate-from-spec.sh 참고.

servers:
  - url: https://coc-api.coc-api.workers.dev
    description: 운영

paths:
  /api/v1/members:
    get:
      operationId: listMembers
      summary: 클랜원 목록
      tags: [members]
      parameters:
        - name: externalId
          in: query
          required: false
          description: CoC 플레이어 태그로 좁힌다. '#'을 포함해 넘긴다
          schema:
            type: string
            example: "#2ABC123"
        - $ref: "#/components/parameters/Includes"
      responses:
        "200":
          description: 클랜원 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListResponse"

  /api/v1/members/{memberId}:
    parameters:
      - name: memberId
        in: path
        required: true
        description: 우리 식별자. CoC 태그가 아니다
        schema:
          type: string
    get:
      operationId: getMember
      summary: 클랜원 한 명
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/Includes"
      responses:
        "200":
          description: 클랜원
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: 그 식별자를 가진 클랜원이 없다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateMember
      summary: 클랜원의 우리 값 수정
      description: |
        우리가 정하는 값만 고친다. 이름·직책·트로피처럼 CoC 가 주인인 값은
        여기서 고칠 수 없다. 고쳐 봐야 다음 동기화가 되돌린다.

        보낸 필드만 바뀐다. 보내지 않은 필드는 그대로 두고, null 을 보내면
        비운다.
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: 고친 뒤의 클랜원
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: |
            본문이 규칙에 맞지 않다. 고칠 값을 하나도 보내지 않았거나,
            비울 수 없는 값(warnings)에 null 을 보냈다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 그 식별자를 가진 클랜원이 없다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/clans:
    get:
      operationId: listClans
      summary: 클랜 목록
      description: |
        우리가 다루는 클랜. 지금은 미니언즈 하나뿐이다.

        마크·레벨·점수·전적은 담지 않는다. CoC 가 주인인 값이라 우리가
        사본을 들고 있으면 두 곳에서 관리하게 되고 언젠가 어긋난다.
        필요하면 external_id 로 CoC 에 물으면 된다.
      tags: [clans]
      responses:
        "200":
          description: 클랜 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClanListResponse"

components:
  parameters:
    Includes:
      name: include
      in: query
      required: false
      description: |
        기본 응답에 함께 실을 덩어리. 쉼표로 여럿 적는다.

        기본 응답은 우리 DB 한 행만 읽는다. CoC 를 부르지도, 무엇을 계산하지도
        않는다. 명단만 필요한 요청에 그 비용을 얹지 않으려는 것이다.

        부르지 않은 덩어리는 키 자체가 없다. null 로 채우지 않는다. 묻지 않은
        것과 값이 없는 것은 다르기 때문이다.

          profile  CoC 가 주인인 값 — 이름·직책·홀·트로피·기부
          league   리그전 선발 — 이번 달 점수·순위·등급
      schema:
        type: array
        items:
          $ref: "#/components/schemas/MemberInclude"
      style: form
      explode: false
      example: profile

  schemas:
    ClanStatus:
      type: string
      description: |
        우리가 이 클랜을 다루는가. CoC 가 주지 않고 우리가 판정한다.
        MemberStatus 와 값이 같지만 뜻이 다르다. 이쪽은 "우리 관심사인가"이고
        저쪽은 "클랜에 남아 있는가"다. 한쪽만 값이 늘 수 있으므로 따로 둔다.
      enum: [ACTIVE, INACTIVE]

    Clan:
      type: object
      description: |
        CoC 가 모르는 값만 담는다. 마크·레벨·점수·전적은 CoC 가 주인이라
        우리가 사본을 들면 두 곳에서 관리하게 된다.

        id 는 우리 식별자이고 externalId 는 CoC 세계의 식별자다. 주소에서
        '#' 을 인코딩하지 않으려고 따로 둔다.

        이름조차 우리가 관리한다. 게임에서 클랜 이름을 바꿔도 우리 표기는
        그대로 둘 수 있고, 태그는 바뀌지 않으므로 그것으로 찾는다.
      required:
        - id
        - externalId
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: 우리 식별자(UUID 문자열)
          example: "3f2a1b4c-5d6e-4f70-8a91-b2c3d4e5f607"
        externalId:
          type: string
          description: CoC 클랜 태그. 이 값으로 CoC API 를 부른다
          example: "#2C8L822LQ"
        displayName:
          type: string
          nullable: true
          description: 우리가 붙이는 이름. 첫 동기화 때 채운다
          example: "미니언즈"
        status:
          $ref: "#/components/schemas/ClanStatus"
        createdAt:
          type: string
          description: 처음 본 시각. ISO 8601(UTC)
          example: "2026-09-11T07:18:57Z"
        updatedAt:
          type: string
          description: 마지막으로 갱신한 시각. ISO 8601(UTC)
          example: "2026-09-11T07:18:57Z"

    ClanListResponse:
      type: object
      required: [clans]
      properties:
        clans:
          type: array
          items:
            $ref: "#/components/schemas/Clan"

    ClanRole:
      type: string
      description: |
        게임 안 직책. CoC API 표기를 대문자로 바꾼 값이다.
        ADMIN 은 게임 화면에서 "장로(Elder)"로 불린다.
        UNKNOWN 은 CoC 가 우리가 모르는 값을 보냈다는 뜻이다.
      enum: [LEADER, COLEADER, ADMIN, MEMBER, UNKNOWN]

    MemberStatus:
      type: string
      description: |
        클랜 소속 여부. CoC API 가 주지 않고 우리가 판정한다.
        목록에 있으면 ACTIVE, 사라지면 INACTIVE 로 내린다.
      enum: [ACTIVE, INACTIVE]

    MemberInclude:
      type: string
      description: |
        includes 로 고를 수 있는 덩어리 이름.

        어휘는 둘로 정해져 있다.

          profile  CoC 가 주인인 값 — 이름·직책·홀·트로피·기부
          league   리그전 선발 — 이번 달 점수·순위·등급

        league 는 늘 이번 달을 가리킨다. 지난 달을 고르는 방법은 아직 정하지
        않았고, 응답에 month 를 함께 실어 어느 달 값인지 밝힌다.

        league 는 아직 열거형에 없다. 그달 점수가 있어야 나오는데 클랜전 기록이
        아직 들어오지 않았다. 그 티켓에서 더한다. 없는 것을 미리 적어 두면
        부르는 쪽이 있다고 믿는다.
      enum: [profile]

    Member:
      type: object
      description: |
        클랜원 한 명. 게임 계정 하나를 가리키며 사람이 아니다.

        여기 담는 것은 우리가 주인인 값뿐이다. DB 한 행을 읽으면 나오고, CoC 를
        부르지도 무엇을 계산하지도 않는다. 명단만 필요한 요청에 그 비용을 얹지
        않으려는 것이다.

        id 는 우리 식별자이고 externalId 는 CoC 세계의 식별자다. 주소에서
        '#' 을 인코딩하지 않으려고 따로 둔다. Clan 과 같은 규칙이다.

        이름과 직책은 여기 없다. CoC 가 주인인 값이라 profile 로 내렸다.
        등급도 여기 없다. 이번 달 점수가 정하는 값이라 league 가 맡는다.

        displayName 은 사람이 정한 표기이며 동기화가 채우지 않는다. 채우면 그
        값이 사람이 정한 것인지 동기화가 써 넣은 것인지 구분할 수 없다. 비었을
        때 무엇을 보여줄지는 화면이 정한다 — profile.name 으로 대신하는 것이
        보통이다.
      required:
        - id
        - externalId
        - status
        - warnings
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: 우리 식별자(UUID 문자열)
          example: "3f2a1b4c-5d6e-4f70-8a91-b2c3d4e5f607"
        externalId:
          type: string
          description: CoC 플레이어 태그. 이 값으로 CoC API 를 부른다
          example: "#2ABC123"
        displayName:
          type: string
          nullable: true
          description: 사람이 정한 표기. 아무도 고치지 않았으면 null 이다
          example: "히로형"
        status:
          $ref: "#/components/schemas/MemberStatus"
        warnings:
          type: integer
          description: 경고 횟수. 표시만 하고 점수에 영향을 주지 않는다
          default: 0
        description:
          type: string
          nullable: true
          description: 관리자 메모. 동기화가 덮어쓰지 않는다
        createdAt:
          type: string
          description: 처음 본 시각. ISO 8601(UTC)
          example: "2026-09-11T07:18:57Z"
        updatedAt:
          type: string
          description: 마지막으로 바뀐 시각. CoC 에서 받은 시각은 profile.fetchedAt 이다
          example: "2026-09-11T07:18:57Z"
        profile:
          $ref: "#/components/schemas/MemberProfile"

    MemberProfile:
      type: object
      description: |
        CoC 가 주인인 값. includes=profile 로 부를 때만 실린다.

        지금은 동기화가 받아 둔 사본을 읽는다. 앞으로 사본을 걷어내면 이 덩어리가
        CoC 호출이 된다. 그때 부르는 쪽을 고치지 않아도 되도록 지금 갈라 둔다.

        fetchedAt 이 이 값들의 나이를 말해 준다. 30분마다 도는 동기화가 채우며,
        아직 한 번도 동기화하지 않았으면 null 이다.
      required: [name, role]
      properties:
        name:
          type: string
        role:
          $ref: "#/components/schemas/ClanRole"
        townhall:
          type: integer
          nullable: true
        trophies:
          type: integer
          nullable: true
        donations:
          type: integer
          nullable: true
        donationsReceived:
          type: integer
          nullable: true
        fetchedAt:
          type: string
          nullable: true
          description: CoC 가 준 값을 받아 적은 시각. ISO 8601(UTC)
          example: "2026-09-14T09:00:00Z"

    MemberUpdate:
      type: object
      description: |
        클랜원에게서 사람이 정하는 값. PATCH 의 본문이다.

        전부 선택 사항이며, 보낸 것만 바뀐다. 값을 비우려면 null 을 보낸다.
        하나도 보내지 않으면 고칠 것이 없으므로 거절한다.

        등급은 여기 없다. 그달 점수가 정하는 값이라 사람이 고치지 않는다.
      minProperties: 1
      properties:
        displayName:
          type: string
          nullable: true
          description: 사람이 정한 표기. null 을 보내면 지우고 화면은 CoC 이름으로 돌아간다
          example: "히로형"
        warnings:
          type: integer
          minimum: 0
          description: 경고 횟수
        description:
          type: string
          nullable: true
          description: 관리자 메모

    MemberListResponse:
      type: object
      required: [members]
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"

    Error:
      type: object
      description: |
        요청을 받아들이지 못했을 때의 본문. FastAPI 가 HTTPException 을
        내보내는 모양과 같다.
      required: [detail]
      properties:
        detail:
          type: string
          example: "그 클랜원이 없습니다"